private-repos/video-clip-assembler/requirements.txt
google-generativeai>=0.8.6
imageio-ffmpeg>=0.6.0
httpx>=0.28.0
Pillow>=10.0.0
